requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
python-dotenv==1.0.0
google-generativeai==0.3.2
reportlab==4.0.4
html2image==3.1.0
spacy==3.7.2
gtts==2.4.0
twilio==8.10.0
bcrypt==4.0.1 